+++ ios/Plugin/VideoPicker/VideoPickerViewController.swift
@@ -28,11 +28,17 @@ extension VideoPickerViewController: UIImagePickerControllerDelegate {
                 let mediaType = info[UIImagePickerController.InfoKey.mediaType] as? String,
                 mediaType == (kUTTypeMovie as String),
                 let url = info[UIImagePickerController.InfoKey.mediaURL] as? URL
-              else { return }
+            else {
+                return
+            }
             let vId: [String: Any] = ["videoUrl": url]
             NotificationCenter.default.post(name: .videoPathInternalReady, object: nil, userInfo: vId)
       }
+    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+        let vId: [String: Any] = ["videoUrl": ""]
+        NotificationCenter.default.post(name: .videoPathInternalReady, object: nil, userInfo: vId)
     }
+}
 
 // MARK: - UINavigationControllerDelegate
 extension VideoPickerViewController: UINavigationControllerDelegate {